Implement a really back lock-free queue.
It's really bad because push_back, which is going to be an important
operation, requires iterating to the end of a linked-list every time. It
also has a shared_lock for pop_front and pop_back that should prevent
head from ever being junk.

Also, this thing isn't heavily tested.